+++ docs/ReleaseNotes.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [3.7.2] - 2024.07.27 unpublished
 
+- Improved user prompts for input.
+
 ## [3.7.1] - 2024.07.27
 
 - Minor fixes and code formatting improvements.